dz3 | split index.js
